+++ src/plone/pdfexport/tests/test_view_aspdf.py
@@ -1,10 +1,9 @@
 # -*- coding: utf-8 -*-
 import unittest
 
-from plone.app.contenttypes.behaviors.leadimage import ILeadImage
+from bs4 import BeautifulSoup
 from plone.app.testing import TEST_USER_ID, setRoles
 from zope.component import getMultiAdapter, queryMultiAdapter
-from zope.component.interfaces import ComponentLookupError
 from zope.viewlet.interfaces import IViewletManager
 
 from plone import api
@@ -58,6 +57,96 @@ class ViewsIntegrationTest(unittest.TestCase):
             v for v in manager.viewlets if v.__name__ == "contentleadimage"
         ]
         self.assertEqual(len(leadimage_viewlet), 1)
+
+    def test_clean_html_with_body_class(self):
+        view = self.portal["news-item"].restrictedTraverse("@@aspdf")
+        clean_hml_str = view._clean_html(
+            raw_html="""
+<html>
+<head></head>
+<body class="template-test-view col-content">
+<div>Header</div>
+
+<div id="content" >
+<h1>PDF content</h1>
+<p>Lorem eiusmod quis adipisicing do exercitation consequat eiusmod est sunt consequat. Sunt magna mollit qui dolore eiusmod ullamco. Qui minim aliquip aute veniam culpa pariatur pariatur cupidatat voluptate voluptate excepteur tempor. Et ea consequat ad voluptate ipsum minim occaecat. Dolore consequat officia aliqua est voluptate do sint voluptate fugiat duis in do sint. Aute exercitation commodo nisi duis non.
+
+Velit laborum exercitation anim sit sunt irure officia ea qui nisi consequat. Magna qui qui commodo nulla laborum non. Nulla sunt amet amet eu magna duis magna sit sint. Magna deserunt reprehenderit qui nostrud fugiat dolor sunt. Commodo nisi ut qui duis eu nisi Lorem sunt. Id adipisicing elit sit in officia veniam eiusmod. Commodo dolore amet deserunt et dolore enim ipsum velit aliquip irure nostrud magna magna veniam.
+
+Fugiat voluptate anim nulla eiusmod sit. Eiusmod irure id voluptate labore. Lorem officia pariatur sunt culpa esse aute anim quis ad enim non nisi Lorem. Cillum duis magna in adipisicing eiusmod tempor tempor labore ullamco excepteur ad officia velit. Deserunt aute ex duis sit et cillum. Tempor irure do eu et sint sint incididunt cupidatat adipisicing fugiat incididunt.</p>
+</div>
+
+</body>
+</html>
+        """
+        )
+        self.assertTrue(clean_hml_str)
+        soup = BeautifulSoup(clean_hml_str, "html.parser")
+        content = soup.select_one("#content")
+        self.assertEqual(
+            content.get("class"),
+            ["template-test-view", "col-content"],
+        )
+
+    def test_clean_html_with_body_class_preserve_content_classes(self):
+        view = self.portal["news-item"].restrictedTraverse("@@aspdf")
+        clean_hml_str = view._clean_html(
+            raw_html="""
+<html>
+<head></head>
+<body class="template-test-view col-content">
+<div>Header</div>
+
+<div id="content" class="content-class">
+<h1>PDF content</h1>
+<p>Lorem eiusmod quis adipisicing do exercitation consequat eiusmod est sunt consequat. Sunt magna mollit qui dolore eiusmod ullamco. Qui minim aliquip aute veniam culpa pariatur pariatur cupidatat voluptate voluptate excepteur tempor. Et ea consequat ad voluptate ipsum minim occaecat. Dolore consequat officia aliqua est voluptate do sint voluptate fugiat duis in do sint. Aute exercitation commodo nisi duis non.
+
+Velit laborum exercitation anim sit sunt irure officia ea qui nisi consequat. Magna qui qui commodo nulla laborum non. Nulla sunt amet amet eu magna duis magna sit sint. Magna deserunt reprehenderit qui nostrud fugiat dolor sunt. Commodo nisi ut qui duis eu nisi Lorem sunt. Id adipisicing elit sit in officia veniam eiusmod. Commodo dolore amet deserunt et dolore enim ipsum velit aliquip irure nostrud magna magna veniam.
+
+Fugiat voluptate anim nulla eiusmod sit. Eiusmod irure id voluptate labore. Lorem officia pariatur sunt culpa esse aute anim quis ad enim non nisi Lorem. Cillum duis magna in adipisicing eiusmod tempor tempor labore ullamco excepteur ad officia velit. Deserunt aute ex duis sit et cillum. Tempor irure do eu et sint sint incididunt cupidatat adipisicing fugiat incididunt.</p>
+</div>
+
+</body>
+</html>
+        """
+        )
+        self.assertTrue(clean_hml_str)
+        soup = BeautifulSoup(clean_hml_str, "html.parser")
+        content = soup.select_one("#content")
+        self.assertEqual(
+            content.get("class"),
+            ["template-test-view", "col-content", "content-class"],
+        )
+
+    def test_clean_html_no_body_class(self):
+        view = self.portal["news-item"].restrictedTraverse("@@aspdf")
+        clean_hml_str = view._clean_html(
+            raw_html="""
+<html>
+<head></head>
+<body>
+<div>Header</div>
+
+<div id="content">
+<h1>PDF content</h1>
+<p>Lorem eiusmod quis adipisicing do exercitation consequat eiusmod est sunt consequat. Sunt magna mollit qui dolore eiusmod ullamco. Qui minim aliquip aute veniam culpa pariatur pariatur cupidatat voluptate voluptate excepteur tempor. Et ea consequat ad voluptate ipsum minim occaecat. Dolore consequat officia aliqua est voluptate do sint voluptate fugiat duis in do sint. Aute exercitation commodo nisi duis non.
+
+Velit laborum exercitation anim sit sunt irure officia ea qui nisi consequat. Magna qui qui commodo nulla laborum non. Nulla sunt amet amet eu magna duis magna sit sint. Magna deserunt reprehenderit qui nostrud fugiat dolor sunt. Commodo nisi ut qui duis eu nisi Lorem sunt. Id adipisicing elit sit in officia veniam eiusmod. Commodo dolore amet deserunt et dolore enim ipsum velit aliquip irure nostrud magna magna veniam.
+
+Fugiat voluptate anim nulla eiusmod sit. Eiusmod irure id voluptate labore. Lorem officia pariatur sunt culpa esse aute anim quis ad enim non nisi Lorem. Cillum duis magna in adipisicing eiusmod tempor tempor labore ullamco excepteur ad officia velit. Deserunt aute ex duis sit et cillum. Tempor irure do eu et sint sint incididunt cupidatat adipisicing fugiat incididunt.</p>
+</div>
+
+</body>
+</html>
+        """
+        )
+        self.assertTrue(clean_hml_str)
+        soup = BeautifulSoup(clean_hml_str, "html.parser")
+        content = soup.select_one("#content")
+        self.assertEqual(
+            content.get("class", []),
+            [],
+        )
 
 
 class ViewsFunctionalTest(unittest.TestCase):